requirements.txt
pycryptodome    